requirements.txt
anyio==3.6.2
APScheduler==3.6.3
attrs==22.2.0
beautifulsoup4==4.11.2
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
decorator==5.1.1
deptry==0.11.0
Flask==2.3.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
MarkupSafe==2.1.3
nltk==3.8.1
pathspec==0.11.1
platformdirs==3.2.0
pycparser==2.21
Pygments==2.14.0
pyperclip==1.8.2
pyTelegramBotAPI==4.7.1
python-dotenv==1.0.0
python-telegram-bot==13.0
pytz==2023.3
requests==2.28.2
requests-cache==0.9.8
requests-file==1.5.1
requests-oauthlib==1.3.1
tornado==6.3.2
tzlocal==5.0.1
urllib3==1.26.16
